rename cosmos -> cosmosdb
